apps/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
redis>=5.2.0
requests>=2.31.0
PyJWT>=2.8.0
pandas>=2.2.0
openpyxl>=3.1.0
celery[redis]>=5.4.0

# Local monorepo package (install with: pip install -e ../../packages/teambition-client)
# teambition-client
